refactor: centralize historical membership persistence
